refactor: move clocks/collection/decorator in their own modules
