+++ tests/test_tick.py
@@ -1,7 +1,9 @@
 import pytest
 
-from ticktock import timer
-from ticktock.timer import Clock, clear_collection, tick, ticktock  # noqa: F401
+from ticktock import collection as collection_mod
+from ticktock.clocks import Clock, tick
+from ticktock.collection import clear_collection
+from ticktock.timers import ticktock
 
 
 def test_tick_simple(fresh_clock_collection):
@@ -30,9 +32,9 @@ def test_tick_clear_default():
     t = tick()
     t.tock()
 
-    assert len(timer._DEFAULT_COLLECTION.clocks) == 1
+    assert len(collection_mod._DEFAULT_COLLECTION.clocks) == 1
     clear_collection()
-    assert len(timer._DEFAULT_COLLECTION.clocks) == 0
+    assert len(collection_mod._DEFAULT_COLLECTION.clocks) == 0
 
 
 @pytest.mark.parametrize("name_tick", [None, "ok"])
