+++ ticktock/__init__.py
@@ -1,10 +1,6 @@
-from ticktock.timer import (  # noqa: F401
-    disable,
-    enable,
-    set_format,
-    set_period,
-    tick,
-    ticktock,
-)
+# flake8: noqa: F401
+from ticktock.clocks import tick
+from ticktock.collection import disable, enable, set_format, set_period
+from ticktock.timers import ticktock
 
 __version__ = "0.0.7"
